+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.0.2-pre
-*
+* Colourise success and error output
 
 ### 0.0.1
 * Initial version, extracted from cliffano/bagofholding